The Development Macros import file.
Includes PROCESS_CLASSY_REPORTS (the main purpose of this repository) plus a couple of smaller, useful macros:

BASIC_HEADER_SETUP: I have this tied to a shortcut of Ctrl-J. It just performs a quick, basic formatting of your header row for most any CSV you might download from any payment processor servicer. It boldfaces the header row text, colors the background yellow, and draws a cell border. It's basic but useful.

PREP_FILE_FOR_CSV_SAVE: When you have an Excel-formatted file (.xlsx or .xls, for example) and you want to save it as CSV, you need to get rid of any "extraneous" rows or columns in your worksheet, or otherwise you'll create an invalid CSV file with a bunch of extra, unwanted commas and so forth. **NOTE that this macro ASSUME that your data begins at cell A1 (that cell A1 is not blank) and is continuous (no skipped rows or columns)! It assume that EVERY used row has the same number of fields and that ALL fields are used.**

I guess at some point I could write a version of this utility that allows some blank cells in a row, but I've never had a need for it. Almost always when I want to create a CSV, for my purposes it's REQUIRED that every row have the same number of fields (cells) and that NO cell be blank. And I always start at A1 (and usually have a header row on row A, but that's not required for this macro).