New Instructions and XLSM file
The instructions have been moved out of the XLSM file and into their own PDF. These instructions now also include the possibility of users using a URL to automatically create their details reports in My Reports.

The XLSM file is now the only distribution method for the code. Previously I also supplied a .BAS file, but the .BAS modules are already embedded in the XLSM file, so that's just extra work. Also, this version of XLSM file is much improved. There are now two modules -- LIBRARY and CLASSY -- instead of the one (Devo_Macros). This separation of concerns allowed me to remove one extraneous macro (the basic header formatter) that was in the previous release.